added framework to dodge bombs

+++ agent_code/foxdevilswild/callbacks.py
@@ -3,21 +3,17 @@ import pickle
 import random
 
 import numpy as np
-
-
+from itertools import  product
 ACTIONS = ['UP', 'RIGHT', 'DOWN', 'LEFT', 'WAIT', 'BOMB']
 norm = lambda x: np.sqrt(x[0]**2+x[1]**2)
-discretized_winkels = np.linspace(-8, 8, num=17)
-discretized_radii = range(0, 22)
+discretized_angles =np.arange(17)
 
 # zustandsdict
 zustandsdict = {}
 counter = 0
 
-for mean_winkel in discretized_winkels:
-    zustandsdict[counter] = (mean_winkel, 0)
-    zustandsdict[counter+17] = (mean_winkel, 1)
-    zustandsdict[counter+2*17] = (mean_winkel, 2)
+for angle, borderstatus, bombstatus in product(discretized_angles,[0,1,2],[0,1]):
+    zustandsdict[counter] = (angle, borderstatus, bombstatus)
     counter += 1
 
 reverse_dict = {v: k for k, v in zustandsdict.items()}
@@ -38,10 +34,9 @@ def setup(self):
     :param self: This object is passed to all callbacks and you can set arbitrary values.
     """
 
-    if self.train or not os.path.isfile("my-saved-model.pt"):
+    if (self.train or not os.path.isfile("my-saved-model.pt")) and 0==1:
         self.logger.info("Setting up model from scratch.")
-        weights = np.random.rand(len(ACTIONS))
-        self.Q = np.zeros((17 * 3, 6))
+        self.Q = np.random.rand(17 * 3*2, 6)*100
         with open("my-saved-model.pt", "wb") as file:
             pickle.dump(self.Q, file)
     else:
@@ -65,11 +60,11 @@ def act(self, game_state: dict) -> str:
     status = state_to_features(game_state)
 
     # todo Exploration vs exploitation
-    random_prob = .2
+    random_prob = 0.1
     if self.train and random.random() < random_prob:
         self.logger.debug("Choosing action purely at random.")
         # 80%: walk in any direction. 10% wait. 10% bomb.
-        return np.random.choice(ACTIONS, p=[.2, .2, .2, .2, .2, .0])
+        return np.random.choice(ACTIONS, p=[.2, .2, .2, .2, .1, .1])
 
     self.logger.debug("Querying model for action.")
 
@@ -96,25 +91,34 @@ def state_to_features(game_state: dict) -> np.array:
     :param game_state:  A dictionary describing the current game board.
     :return: np.array
     """
-
+    own_position = np.array(game_state['self'][-1])
+    #the following block of code refers to the state property of coin finding
     coin_positions      = game_state['coins']
-    own_position       = np.array(game_state['self'][-1])
-    relative_coin_positions = [(item[0]- own_position[0],item[1] - own_position[1]) for item in coin_positions]
-    radii = [norm(x) for x in relative_coin_positions]
-    minimal_radius = np.argmin(radii)
+    if any(coin_positions):
+        relative_coin_positions = coin_positions-own_position
+        radii =  np.apply_along_axis(norm,1,relative_coin_positions)
+        minimal_radius = np.argmin(radii)
+        # Calculate angle w.r.t nearest coin
+        x = relative_coin_positions[minimal_radius][0]
+        y = relative_coin_positions[minimal_radius][1]
+        angle = np.arctan2(x, y)+np.pi
+        discretized_angle = int(angle *2.545)
+        #print(discretized_angle)
+    else:
+        discretized_angle = 16
 
-    # Calculate angle w.r.t nearest coin
-    x = relative_coin_positions[minimal_radius][0]
-    y = relative_coin_positions[minimal_radius][1]
-    angle = np.arctan2(x, y)
-
-
-    discretized_winkel = int(angle *2.86)
-
+    #the following block of code refers to the state property of invalid movements
     borderstatus = 0
     if not game_state["field"][tuple(own_position-[0,1])] == 0:# or game_state["field"][tuple(own_position-[0,1])] == 1:
         borderstatus =1
     if not game_state["field"][tuple(own_position-[1,0])] == 0:# or game_state["field"][tuple(own_position-[1,0])] == 1:
         borderstatus =2
 
-    return reverse_dict[(discretized_winkel, borderstatus)]
+    #the following block of code refers to the state property of placing bombs
+    bombstatus = 0
+    bombs = np.array(sum(np.array(game_state["bombs"]).reshape((len(game_state["bombs"]), 2))[:, 0], ())).reshape(len(game_state["bombs"]),2)
+
+    if bombs.any() and (own_position-bombs).flatten().any() and np.apply_along_axis(norm,1,own_position - bombs).any()<4:
+        bombstatus = 1
+    #print("state: ", reverse_dict[(discretized_angle, borderstatus, bombstatus)])
+    return reverse_dict[(discretized_angle, borderstatus, bombstatus)]
